Reworked how systems are stored and fetched

+++ frontend/src/components/UI/DeleteBar/DeleteBar.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import Snackbar from '@material-ui/core/Snackbar';
+import SnackbarContent from '@material-ui/core/SnackbarContent';
 import IconButton from '@material-ui/core/IconButton';
 import CheckIcon from '@material-ui/icons/Check';
 import ClearIcon from '@material-ui/icons/Clear';
@@ -7,8 +8,9 @@ import { makeStyles } from '@material-ui/core/styles';
 // Redux
 import { connect } from 'react-redux';
 import { getJournalByObjectBySystem } from '../../../store/actions/cable';
-import { cableDeleteRemoveAll, cableDeleteSelected } from '../../../store/actions/delete';
-import { undoClear, undoCableJournalDataSave } from '../../../store/actions/undo';
+import { cableDeleteRemoveAll, cableDeleteSelected, 
+  estimateDeleteRemoveAll, estimateDeleteSelected } from '../../../store/actions/delete';
+import { undoClear, undoCableJournalDataSave, undoEstimateDataSave } from '../../../store/actions/undo';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -17,11 +19,15 @@ const useStyles = makeStyles((theme) => ({
       marginTop: theme.spacing(2),
     },
   },
+  snack: {
+    color: 'white',
+    backgroundColor: 'red',
+  },
 }));
 
 const DeleteBar = (props) => {
   const classes = useStyles();
-  const { deleteEnabled, deleteItemsNumber, deleteData, deleteType } = props;
+  const { deleteEnabled, deleteItemsNumber, deleteData, deleteType, deleteSelectorEnabled } = props;
   
   // Clicking delete button
   const deleteClickHandler = () => {
@@ -29,6 +35,10 @@ const DeleteBar = (props) => {
       case 'cable_journal':
         props.cableDeleteSelected(deleteData);
         props.undoCableJournalDataSave();
+        break;
+      case 'estimates':
+        props.estimateDeleteSelected(deleteData);
+        props.undoEstimateDataSave();
         break;
       default:
         break;
@@ -41,18 +51,23 @@ const DeleteBar = (props) => {
         props.cableDeleteRemoveAll();
         props.undoClear();
         break;
+      case 'estimates':
+        props.estimateDeleteRemoveAll();
+        props.undoClear();
+        break;
       default:
         break;
     }
   };
-
-  return (
-  <div className={classes.root}>
-  <Snackbar
-  open={deleteEnabled}
-  message={`Удалить выбранные элементы? (${deleteItemsNumber} шт.)`}
-  key="delsnack"
-  action={
+  let deleteSnack = null;
+  if (deleteSelectorEnabled) {
+    deleteSnack = (
+      <div className={classes.root}>
+    <Snackbar open={deleteEnabled}>
+    <SnackbarContent className={classes.snack}
+    message={`Удалить выбранные элементы? (${deleteItemsNumber} шт.)`}
+    key="delsnack"
+    action={
     <React.Fragment>
       <IconButton onClick={() => deleteClickHandler()}>
         <CheckIcon style={{ color: 'white'}}/>
@@ -60,11 +75,14 @@ const DeleteBar = (props) => {
       <IconButton onClick={() => cancelClickHandler()}>
         <ClearIcon style={{ color: 'white'}}/>
       </IconButton>
-    </React.Fragment>
-  }/>
+    </React.Fragment> }/>
+  </Snackbar>
   </div>
-  );
-} 
+    );
+  };
+
+  return deleteSnack;
+};
 
 const mapStateToProps = state => {
   return {
@@ -72,9 +90,11 @@ const mapStateToProps = state => {
     deleteData: state.del.deleteData,
     deleteType: state.del.deleteType,
     deleteItemsNumber: state.del.deleteItemsNumber,
+    deleteSelectorEnabled: state.sel.deleteSelectorEnabled,
   };
 };
 
 export default connect(mapStateToProps, 
   { cableDeleteRemoveAll, cableDeleteSelected, getJournalByObjectBySystem,
-  undoClear, undoCableJournalDataSave })(DeleteBar);
+    estimateDeleteRemoveAll, estimateDeleteSelected,
+  undoClear, undoCableJournalDataSave, undoEstimateDataSave })(DeleteBar);
